Updated folder name and finalized code
Running the main sends the 10 outputs instances to the csv file
