+++ Genetic Programming/GP.csv
@@ -1,0 +1,11 @@
+Run,Training Duration (ms),Training Accuracy (%),Test Accuracy (%),Seed,Correct/Total
+1,1006,93.19,99.62,1747850879782,262/263
+2,731,93.19,100.00,1747850880803,263/263
+3,543,93.19,100.00,1747850881534,263/263
+4,553,93.19,100.00,1747850882077,263/263
+5,779,93.19,99.24,1747850882630,261/263
+6,836,93.19,100.00,1747850883410,263/263
+7,565,94.79,89.73,1747850884246,236/263
+8,728,85.87,84.41,1747850884812,222/263
+9,739,93.19,100.00,1747850885540,263/263
+10,642,93.19,99.24,1747850886279,261/263
