+++ Genetic Programming/Node.java
